requirements.txt
numpy
simpleitk
matplotlib
plotly
mpld3
fastapi
uvicorn
pydantic==1.10.12
pydantic[dotenv]
azure-storage-blob